# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: UNIGE Events API
  version: 1.0.0
  description: |
    API REST du backend UNIGE Events (Java 21 / Quarkus 3.32).
    Source de vérité : unige-events-api/docs/openapi/openapi.yaml
    Ce fichier est synchronisé avec le repo frontend. Mettre à jour les deux repos ensemble.
    Tous les champs sont en camelCase. Les booléens n'ont pas de préfixe is.
servers:
  - url: /api
    description: Proxié par Nginx (prod) ou Vite (dev)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT   # Auth0 / OIDC — validé par quarkus-oidc
  headers:
    RetryAfter:
      description: |
        Nombre de secondes à attendre avant de retenter la requête (cf. RFC 6585 §4).
        Présent sur les réponses 429 émises par les endpoints de mutation rate-limités
        (voir ticket SECURITY-pentest 4.14). La valeur est toujours **>= 1** :
        l'implémentation arrondit le reste de la fenêtre vers le haut au moins à 1
        seconde — un client qui retente immédiatement n'a aucune chance d'être admis.
      required: true
      schema:
        type: integer
        minimum: 1
      example: 45
  responses:
    RateLimited:
      description: |
        Trop de requêtes — la limite par utilisateur sur cet endpoint a été dépassée
        (cf. pentest 2026-04-17 finding 4.14, ticket #98). L'en-tête `Retry-After`
        indique le nombre de secondes à attendre avant de retenter.
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            error: rate_limited
            message: "Too many requests. Retry in 45 seconds."
  schemas:
...
    Notification:
      type: object
      description: |
        Notification in-app retournée par `GET /api/users/me/notifications`. Owned par
        `notification-service`. Le payload est composé par les consumers Kafka du service
        à partir des événements `events.cancelled`, `events.updated` et `attendances.created`
        (cf. `backend/docs/architecture.md` section Kafka). Les notifs `EVENT_REMINDER` sont
        réservées pour un scheduler ultérieur (hors scope SCRUM-99).

        **Extensibilité (phase 2).** Les valeurs `NEW_FOLLOWER`, `FOLLOW_REQUEST`,
        `FOLLOW_ACCEPTED` (SCRUM-140), `COMMENT_MENTION` et `NEW_COMMENT` (SCRUM-145)
        seront ajoutées à l'enum `type` ainsi qu'à la CHECK constraint DB via
        `V2__widen_notification_type_check.sql` lors de la livraison de la phase 2.
        Pour `NEW_FOLLOWER` / `FOLLOW_REQUEST` / `FOLLOW_ACCEPTED`, le champ
        `relatedUserId` portera l'UUID du follower / requester / accepter.
      required: [id, userId, type, message, read, createdAt]
      properties:
        id:
          type: integer
          format: int64
          description: Identifiant numérique stable (clé primaire DB).
        userId:
          type: string
          format: uuid
          description: UUID du destinataire (`users.id`, sans FK cross-DB).
        type:
          type: string
          enum: [EVENT_UPDATED, EVENT_CANCELLED, EVENT_REMINDER, NEW_ATTENDEE]
          description: |
            Type métier de la notification. Phase 1 : 4 valeurs. Phase 2 : 5 valeurs
            additionnelles (cf. description du schéma).
        eventId:
          type: integer
          format: int64
          nullable: true
          description: |
            ID de l'événement lié (nullable — sera `null` pour les notifs phase 2
            non-event-bound comme `NEW_FOLLOWER`).
        relatedUserId:
          type: string
          format: uuid
          nullable: true
          description: |
            Acteur secondaire de la notification. Phase 1 : UUID de l'inscrit pour
            `NEW_ATTENDEE`, `null` pour `EVENT_UPDATED` / `EVENT_CANCELLED`. Phase 2 :
            UUID du follower pour `NEW_FOLLOWER`, UUID du mentionneur pour
            `COMMENT_MENTION`, etc.
        message:
          type: string
          description: |
            Message textuel pré-composé par le consumer Kafka — le frontend l'affiche
            tel quel. Exemple : `"L'événement « Concert d'ouverture » a été annulé."`.
        read:
          type: boolean
          description: État de lecture. Par défaut `false` à la création, passe à `true`
            via `PATCH /users/me/notifications/{id}/read` ou `read-all`.
        createdAt:
          type: string
          format: date-time
          description: Timestamp serveur à la création (immutable).
        readAt:
          type: string
          format: date-time
          nullable: true
          description: |
            Timestamp à la transition `read=false → true`. `null` tant que la notif
            n'est pas lue ; immuable une fois posé (un re-mark-read est idempotent).
    ReadAllResponse:
      type: object
      description: Body de retour de `PATCH /api/users/me/notifications/read-all`.
      required: [updated]
      properties:
        updated:
          type: integer
          format: int64
          description: Nombre de rows passées de `read=false` à `read=true` par l'appel.
            `0` si toutes les notifs étaient déjà lues (l'appel reste idempotent).
...
    ApiErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
paths:
...
  /users/me/notifications:
    get:
      summary: Liste des notifications de l'utilisateur connecté (non lues en premier)
      description: |
        Retourne les notifications du caller, triées **unread first puis `createdAt DESC,
        id DESC`** (SCRUM-99). Paginées via `?page=0&size=20`. Le header de réponse
        `X-Unread-Count` porte le nombre total de notifs non lues du caller (inconditionnel,
        même `0`), ce qui évite un round-trip dédié pour le badge UI.
      operationId: getMyNotifications
      tags: [notifications]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Liste paginée des notifications
          headers:
            X-Unread-Count:
              description: Nombre total de notifications `read=false` du caller (toutes pages confondues).
              schema:
                type: integer
                format: int64
                minimum: 0
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '401':
          description: Token absent, invalide, ou profil non provisionné côté user-service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /users/me/notifications/{id}/read:
    patch:
      summary: Marquer une notification comme lue
      description: |
        Idempotent — une notif déjà `read=true` renvoie `204` sans modification.
        Anti-oracle : une notification appartenant à un autre user retourne `404`
        (même envelope qu'un id inexistant).
      operationId: markMyNotificationRead
      tags: [notifications]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Notification marquée comme lue (idempotent).
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Notification introuvable, ou appartenant à un autre user (anti-oracle)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /users/me/notifications/read-all:
    patch:
      summary: Marquer toutes les notifications du caller comme lues
      description: |
        Opération bulk. Retourne le nombre de rows passées de `read=false` à `read=true`.
        Idempotent — appel répété sur un fil déjà entièrement lu retourne `{ "updated": 0 }`.
      operationId: markAllMyNotificationsRead
      tags: [notifications]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notifications marquées comme lues (bulk update)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadAllResponse'
        '401':
          description: Token absent ou invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
...